feat: Add ability to create sample app
